+++ tests/scripts/get-remediation-capi.php
@@ -46,6 +46,7 @@ $cacheRedisConfigs = [
 $redisCache = new Redis($cacheRedisConfigs, $logger);
 // Init CAPI remediation
 $remediationConfigs = [];
-$remediationEngine = new CapiRemediation($remediationConfigs, $capiClient, $redisCache, $logger);
+$remediationEngine = new CapiRemediation($remediationConfigs, $capiClient, $phpFileCache, $logger);
 // Determine the remediation for the given IP
-echo $remediationEngine->getIpRemediation($ip) . \PHP_EOL;
+echo 'Remediation: ' . $remediationEngine->getIpRemediation($ip) . \PHP_EOL;
+echo 'Origins count: ' . json_encode($remediationEngine->getOriginsCount()) . \PHP_EOL;